fix: order building for Vaulting in AfterOrder
