+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 9.7.5
+- Fixes an issue, where vaulted payments could not be processed in the after order payment process in certain situations
+
+# 9.7.4
+- Fixes an issue, where input fields of the refund modal could be missing
+
 # 9.7.3
 - Fixes an issue, where card fields were not displayed in the after order payment process (shopware/shopware#7643)
