+++ tests/OrdersApi/Builder/VenmoOrderBuilderTest.php
@@ -1,0 +1,41 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Swag\PayPal\Test\OrdersApi\Builder;
+
+use Shopware\Core\Framework\Log\Package;
+use Swag\PayPal\OrdersApi\Builder\AbstractOrderBuilder;
+use Swag\PayPal\OrdersApi\Builder\Util\AddressProvider;
+use Swag\PayPal\OrdersApi\Builder\VenmoOrderBuilder;
+use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Venmo;
+use Swag\PayPal\Test\OrdersApi\Builder\Trait\VaultableOrderBuildTrait;
+
+/**
+ * @internal
+ */
+#[Package('checkout')]
+class VenmoOrderBuilderTest extends AbstractOrderBuilderTest
+{
+    use VaultableOrderBuildTrait;
+
+    protected function getBuilder(): AbstractOrderBuilder
+    {
+        return new VenmoOrderBuilder(
+            $this->systemConfig,
+            $this->purchaseUnitProvider,
+            new AddressProvider(),
+            $this->localeCodeProvider,
+            $this->itemListProvider,
+            $this->vaultTokenService,
+        );
+    }
+
+    protected function getPaymentSourceClass(): string
+    {
+        return Venmo::class;
+    }
+}
